pyproject.toml
[tool.poetry]
name = "horrorscopes"
version = "0.1.0"
description = ""
authors = ["Zachary"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
